+++ primitives/manta/src/lib.rs
@@ -27,4 +27,5 @@ pub mod assets;
 pub mod constants;
 pub mod types;
 pub mod util;
+///
 pub mod xcm;